Fix webhook CA when using certmanager
The cert manager implementation of the upstream helm chart has a bug that
it won't update the CA in the webhook configuration once the CA is
reissued, which it will be by default.

This PR adds the `inject-ca-from` annotation which will make sure that
the cert manager will update the CA of the certificate once it is
updated.