Update to latest helm chart version
This will update the component to the latest helm chart.
The latest helm chart contains a few relevant changes:
* The image registry can now be properly set for every image using the
`containerRegistry` parameter in the helm chart. This makes the
post-processing fix obsolete.
* The registry has been changed from `docker.io` to `quay.io` on the
upstream helm chart
* The image to use for setting the RestApi password has been changed
from the `ongres/kubectl` image to the `bitnami/kubectl` image. The
reason is, that we can leverage the `kubernetesVersion` parameter in the
component to fetch the correct `kubectl` container to match the clusters
version.

Signed-off-by: Nicolas Bigler <nicolas.bigler@vshn.ch>